@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-  compileSdkVersion 25
+  compileSdkVersion 29
   buildToolsVersion "25.0.2"
   defaultConfig {
     applicationId "com.xiaopo.flying.stickerview"
-    minSdkVersion 14
-    targetSdkVersion 25
+    minSdkVersion 23
+    targetSdkVersion 29
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
   }
   buildTypes {
     release {
@@ -21,10 +21,10 @@ android {
 
 dependencies {
   compile fileTree(dir: 'libs', include: ['*.jar'])
-  androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+  androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
     exclude group: 'com.android.support', module: 'support-annotations'
   })
-  compile "com.android.support:appcompat-v7:$rootProject.ext.supportLibraryVersion"
+  compile 'androidx.appcompat:appcompat:1.0.0'
   compile "com.github.chrisbanes:PhotoView:$rootProject.ext.photoViewVersion"
   testCompile 'junit:junit:4.12'
   compile project(path: ':sticker')
